+++ core/new-gui/src/app/app.module.ts
@@ -94,8 +94,8 @@ import { ResultTableFrameComponent } from './workspace/component/result-panel/re
 import { DynamicModule } from 'ng-dynamic-component';
 import { RowModalComponent } from './workspace/component/result-panel/result-panel-modal.component';
 import { MonacoEditorModule } from 'ngx-monaco-editor';
-import { FormlyFormFrameComponent } from './workspace/component/property-editor/formly-form-frame/formly-form-frame/formly-form-frame.component';
-import { BreakpointFrameComponent } from './workspace/component/property-editor/breakpoint-frame/breakpoint-frame/breakpoint-frame.component';
+import { OperatorPropertyEditFrameComponent } from './workspace/component/property-editor/operator-property-edit-frame/operator-property-edit-frame.component';
+import { BreakpointPropertyEditFrameComponent } from './workspace/component/property-editor/breakpoint-property-edit-frame/breakpoint-property-edit-frame.component';
 
 registerLocaleData(en);
 
@@ -145,8 +145,8 @@ registerLocaleData(en);
     NgbdModalUserFileShareAccessComponent,
     ConsoleFrameComponent,
     ResultTableFrameComponent,
-    FormlyFormFrameComponent,
-    BreakpointFrameComponent
+    OperatorPropertyEditFrameComponent,
+    BreakpointPropertyEditFrameComponent
   ],
   imports: [
     BrowserModule,
