+++ core/new-gui/src/app/workspace/component/property-editor/breakpoint-property-edit-frame/breakpoint-property-edit-frame.component.html
